+++ Changelog.md
@@ -1,7 +1,10 @@
 # Change Log
 
+## 3.2.0
+- Adding support for Rules API, see [rules documentation](./docs/rules.md) for details
+
 ## 3.1.4
-- Fixing issues with ListType and using Objects for LightState, Issue #134 
+- Fixing issues with ListType and using Objects for LightState, Issue #134
 
 ## 3.1.3
 - Fixing error with setting `transitiontime` in light state, Issue #133